Need to reduce output, it is becoming too big.

+++ java/src/test/java/sample/SampleTest.java
@@ -2,7 +2,6 @@ package sample;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-
 import static sample.Constants.*;
 
 import org.junit.jupiter.api.Test;
@@ -19,14 +18,16 @@ import java.util.List;
 
 class SampleTest {
 
+    final boolean COVERAGE_MODE = false;
+
     @Test
     void combinationTest() throws IllegalAccessException, IOException {
         // all values
-        final List<Integer> allAnsprAufO = Arrays.asList(0);
+        final List<Integer> allAnsprAufO = Arrays.asList(-1, 0, 1); // copied to AnsprAufV
         final List<Integer> allAnsprAufV = Arrays.asList(-1, 0, 1);
-        final List<Integer> allAnsprBand = Arrays.asList(0);
+        final List<Integer> allAnsprBand = Arrays.asList(-1, 0);
         final List<Integer> allAnsprHyst = Arrays.asList(-1, 0, 1);
-        final List<Integer> allAnsprZuO = Arrays.asList(0);
+        final List<Integer> allAnsprZuO = Arrays.asList(0, 1);
         final List<Integer> allAnsprZuV = Arrays.asList(0);
         final List<Integer> allAutoIbsOk = Arrays.asList(0, C_IBS_OK);
         final List<Integer> allBinSteuer = Arrays.asList(0, BO_REGLER);
@@ -36,9 +37,9 @@ class SampleTest {
         final List<Integer> allRegDiff = Arrays.asList(-1, 0, 1);
         final List<Integer> allRegDiffSch = Arrays.asList(-1, 0, 1, 10);
         final List<Integer> allRegMode = Arrays.asList(0, N_AUTOMATIK, N_VALVE_DIAG);
-        final List<Integer> allSollwertRev = Arrays.asList(0);
+        final List<Integer> allSollwertRev = Arrays.asList(0, 10);
         final List<Integer> allStellFwd = Arrays.asList(0); // is out param
-        final List<Integer> allStellIstRev = Arrays.asList(0);
+        final List<Integer> allStellIstRev = Arrays.asList(0, 50);
         final List<Integer> allWirkFall = Arrays.asList(0, 1);
 
         StringBuilder totalState = new StringBuilder();
@@ -114,12 +115,13 @@ class SampleTest {
         Arrays.sort(fields, Comparator.comparing(Field::getName));
         for (Field field : fields) {
             if (Modifier.isPublic(field.getModifiers())) {
-                state.append(field.getName());
-                state.append(';');
+                //state.append(field.getName());
+                //state.append(';');
                 state.append(field.get(null));
-                state.append('\n');
+                state.append(',');
             }
         }
+        state.append('\n');
         return state.toString();
     }
 
@@ -131,10 +133,10 @@ class SampleTest {
         Files.write(receivedFile, received.getBytes());
 
         Path approvedFile = Paths.get(prefix + "approved" + suffix);
-        // coverage mode
-        // if (!approvedFile.toFile().exists()) {
+
+        if (COVERAGE_MODE || !approvedFile.toFile().exists()) {
             Files.write(approvedFile, received.getBytes());
-        // }
+        }
 
         String approved = new String(Files.readAllBytes(approvedFile));
         assertEquals(approved, received);
